Fix emitVar store instruction emission condition (#184)
Previously we relied on AST node type to determine whether to emit the
store instruction to init the variable or not. which is not accurate
and fails in cases like local assignment in a function `x := y`

The fix is to check if the variable was initialized using a constant
value of not in the OpVariable, and if it's not, we emit the store
instruction with the given initExpr if any.